Export po_compile() (#234)
* Export dev_compile()

This function was previously called `update_mo_files()`. I don't know if `dev_compile()` is better, but I like the idea of using prefixes to distinguish between functions for developers and translators. Let me know if you have a better name, or if you'd prefer me to leave as is.

I also made a few changes to make it more user friendly: it now discovers the package name (if needed), and can lazily recompile just the files that have changed. I also tweaked the messaging so you get one message per translation (which makes it easier to see what the msgfmt messages apply to).

Two questions:

* Is my scoped use of roxygen2 ok? (roxygen2 is designed to work like this so there's no chance of it clobbering your hand written files)

* Do you want me to add tests? If so, I'd probably start by refactoring the code to extract out a data frame (data table?) that contains the po path, mo_path, and language.

* Rename dev_compile() to po_compile()

* Pull out separate po_files() function

* If not lazy, delete .mo files without matching .po

* typo

* simplify helper

* tweaks

* match signature order

* clarify verbose message

* rename po_files --> get_po_metadata

feels less muddled for describing the output (which involves .mo files)

* get_po_metadata() tweaks

* Hoist lazy up into po_compile()
* Use lang-regexp for filtering files
* Record po type

* Message about translation type

* Move po_compile() to own file

* Tweak test check

* Add a test for get_po_metadata()

* Improve variable access

* dup line

* startsWith over grepl

* Use new form of get_desc_data()

Co-authored-by: Michael Chirico <michaelchirico4@gmail.com>